Time: 30 ms (7.54%), Space: 24.1 MB (69.01%) - LeetHub

+++ 2042-maximum-product-difference-between-two-pairs/2042-maximum-product-difference-between-two-pairs.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int maxProductDifference(vector<int>& nums) {
         sort(nums.begin(), nums.end());
-        unsigned short numsSize = nums.size();
+        int numsSize = nums.size();
 
         return (nums[numsSize - 1] * nums[numsSize - 2]) - (nums[0] * nums[1]);
     }
